moved trending and recommend to components from screen
